% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_test.R
\name{clopper_pearson_ci}
\alias{clopper_pearson_ci}
\title{Calculate Clopper-Pearson confidence interval}
\usage{
clopper_pearson_ci(p, n, alpha = 0.05)
}
\arguments{
\item{p}{The observed rate.}

\item{n}{The sample size.}

\item{alpha}{The two-sided significance level. Default is 0.05.}
}
\value{
A dataframe of Clopper-Pearson confidence interval.
}
\description{
This function calculate the Clopper-Pearson confidence interval.
}
\examples{
# example code
clopper_pearson_ci(0:20/20, 20)
}
