% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_test.R
\name{power_exact_prop}
\alias{power_exact_prop}
\title{Calculate the power of a one sample exact test for rate}
\usage{
power_exact_prop(
  n,
  p,
  null_p,
  alpha = 0.05,
  side = c("U", "L", "Both"),
  margin = 0
)
}
\arguments{
\item{n}{A numeric vector of the total sample size, power will be calculated for
each n.}

\item{p}{The rate in alternative hypothesis.}

\item{null_p}{The rate in null hypothesis.}

\item{alpha}{The two-sided significance level. Default is 0.05, which is a 0.025
one-sided test.}

\item{side}{The side of the test, "U" and "L" for one-sided test in upper or
lower direction. "Both" for a two-sided equivalence test. Default is "U".}

\item{margin}{The effect to be tested. Default = 0.

Recommend to adjust \code{null_p} directly and not change \code{margin}.

The hypothesis is constructed as:

For \code{side} = "U"

H0: p - null_p <= margin

H1: p - null_p > margin

For \code{side} = "L"

H0: p - null_p >= margin

H1: p - null_p < margin

For \code{side} = "Both"

H0: p - null_p = margin

H1: p - null_p != margin}
}
\value{
A dataframe of sample size n and power of the test.
}
\description{
This function calculate the power of a one sample exact test for rate.
}
\examples{
# example code
power_exact_prop(671, p = 0.66, null_p = 0.60)

# For multiple n
power_exact_prop(600:700, p = 0.66, null_p = 0.60)
}
