% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_exact_test.R
\name{power_exact_ve}
\alias{power_exact_ve}
\title{Calculate the power of a Vx efficacy trial using exact method with fixed total
case}
\usage{
power_exact_ve(case, r1 = 0.5, ve, alpha = 0.05, incidence0, margin)
}
\arguments{
\item{case}{A numeric vector of the total cases, power will be calculated for
each case.}

\item{r1}{The proportion of sample1 in the total sample. Default is 0.5, which
lead to equal sample size for sample1 and sample2.}

\item{ve}{The assumed vaccine efficacy in alternative hypothesis. e.g., ve = 0.7
means 70\% efficacy (rate ratio = 1 - 0.7 = 0.3).}

\item{alpha}{The two-sided significance level. Default is 0.05, which is a 0.025
one-sided test.}

\item{incidence0}{The incidence rate in sample2(control group), which is used to
convert cases to sample size.}

\item{margin}{The ve margin to be tested.

The hypothesis is constructed as:

H0: ve <= margin

H1: ve > margin}
}
\value{
A dataframe of number of cases, sample size and power of the test.
}
\description{
This function calculate the power of a Vx efficacy trial.
}
\examples{
# example code
power_exact_ve(case = 80, r1 = 0.5, ve = 0.7, incidence0 = 0.025, margin = 0.3)

# For multiple cases
power_exact_ve(case = 50:100, r1 = 0.5, ve = 0.7, incidence0 = 0.025, margin = 0.3)
}
