% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Farrington-Manning.R
\name{power_Farrington_Manning}
\alias{power_Farrington_Manning}
\title{Calculate the power of a Farrington-Manning test.}
\usage{
power_Farrington_Manning(
  n,
  r1 = 0.5,
  p1,
  p2,
  alpha = 0.05,
  side = c("U", "L", "Both"),
  delta
)
}
\arguments{
\item{n}{A numeric vector of the total sample size, power will be calculated for
each n.}

\item{r1}{The proportion of sample1 in the total sample. Default is 0.5, which
lead to equal sample size for sample1 and sample2.}

\item{p1}{The population rate where sample1 comes from.}

\item{p2}{The population rate where sample2 comes from.}

\item{alpha}{The two-sided significance level. Default is 0.05, which is a 0.025
one-sided test.}

\item{side}{The side of the test, "U" and "L" for one-sided test in upper or
lower direction. "Both" for a two-sided test. Default is "U".}

\item{delta}{The effect to be tested (|delta| <= 1).

The hypothesis is constructed as:

For \code{side} = "U"

H0: rate1 - rate2 <= delta

H1: rate1 - rate2 > delta

For \code{side} = "L"

H0: rate1 - rate2 >= delta

H1: rate1 - rate2 < delta

For \code{side} = "Both"

H0: rate1 - rate2 = delta

H1: rate1 - rate2 != delta}
}
\value{
A dataframe of sample size n and power of the test.
}
\description{
This function calculate the power of a Farrington-Manning test using normal
approximation.
}
\examples{
# example code
# FM test for a NI margin of -0.1, H1: p1 - p2 > -0.1
power_Farrington_Manning(n = 500, p1 = 0.8, p2 = 0.8, delta = -0.1)

# For multiple n
power_Farrington_Manning(n = seq(400, 600, 2), p1 = 0.8, p2 = 0.8, delta = -0.1)
}
\references{
Farrington CP, Manning G. Test statistics and sample size formulae
for comparative binomial trials with null hypothesis of non-zero risk difference
or non-unity relative risk. Stat Med. 1990 Dec;9(12):1447-54.
}
