% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_logscale.R
\name{power_ttest_log}
\alias{power_ttest_log}
\title{Calculate the power of a two sample t-test assuming equal variance for log-normal data}
\usage{
power_ttest_log(
  n,
  r1 = 0.5,
  diff_log10 = 0,
  sd_log10,
  alpha = 0.05,
  side = c("U", "L", "Both"),
  margin
)
}
\arguments{
\item{n}{A numeric vector of the total sample size, power will be calculated for
each n.}

\item{r1}{The proportion of sample1 in the total sample. Default is 0.5, which
lead to equal sample size for sample1 and sample2.}

\item{diff_log10}{The true difference of log10 mean.}

\item{sd_log10}{The standard deviation of log10 data.}

\item{alpha}{The two-sided significance level. Default is 0.05, which is a 0.025
one-sided test.}

\item{side}{The side of the test, "U" and "L" for one-sided test in upper or
lower direction. "Both" for a two-sided equivalence test.

Default is "U".}

\item{margin}{The effect to be tested.

The hypothesis is constructed as:

For \code{side} = "U"

H0: geomean1 / geomean2 <= 1 / margin

H1: geomean1 / geomean2 > 1 / margin

For \code{side} = "L"

H0: geomean1 / geomean2 >= 1 / margin

H1: geomean1 / geomean2 < 1 / margin

For \code{side} = "Both"

H0: geomean1 / geomean2 = 1 / margin

H1: geomean1 / geomean2 != 1 / margin}
}
\value{
A dataframe of sample size n and power of the test.
}
\description{
This function calculate the power of a t-test for log-normal data (e.g., titre).
Equal variance is assumed.
}
\examples{
# example code
power_ttest_log(400, diff_log10 = 0, sd_log10 = 0.5, margin = 1.5)

# For multiple n
power_ttest_log(seq(300, 400, 2), diff_log10 = 0, sd_log10 = 0.5, margin = 1.5)
}
