% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson_score_without_CC.R
\name{power_wilson}
\alias{power_wilson}
\title{Calculate the power of a test for rate difference using Wilson score CI}
\usage{
power_wilson(n, r1 = 0.5, p1, p2, alpha = 0.05, delta)
}
\arguments{
\item{n}{A numeric vector of the total sample size, power will be calculated for
each n.}

\item{r1}{The proportion of sample1 in the total sample. Default is 0.5, which
lead to equal sample size for sample1 and sample2.}

\item{p1}{The population rate where sample1 comes from.}

\item{p2}{The population rate where sample2 comes from.}

\item{alpha}{The two-sided significance level. Default is 0.05, which is a 0.025
one-sided test.}

\item{delta}{The effect to be tested (|delta| <= 1).

The hypothesis is constructed as:

H0: rate1 - rate2 <= delta

H1: rate1 - rate2 > delta}
}
\value{
A dataframe of sample size n and power of the test.
}
\description{
This function calculate the power of a test for rate difference using Wilson
score CI without continuity correction. The function enumerate all possible
binary samples of a given sample size. The power is calculated as the average
possibility (PMF of binomial distribution) of achieving positive test over all
samples.
}
\section{Note}{
The test procedure is actually a one-sided test comparing the lower limit of
Wilson score CI of rate difference and delta.
}

\examples{
# example code

power_wilson(n = 792, p1 = 0.76, p2 = 0.76, delta = -0.1)

# For multiple n
power_wilson(n = seq(300, 500, 10), p1 = 0.99, p2 = 0.99, delta = -0.05)
}
\seealso{
\code{\link[=wilson_rd_ci]{wilson_rd_ci()}} for CI of rate difference.
}
