% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson_score_without_CC.R
\name{wilson_ci}
\alias{wilson_ci}
\title{Calculating Wilson score confidence interval without continuity correction}
\usage{
wilson_ci(x, n, alpha = 0.05, check = TRUE)
}
\arguments{
\item{x}{A numeric vector of the numerator of the rate.}

\item{n}{A numeric vector of the denominator of the rate.}

\item{alpha}{The significance level. Default is 0.05 which gives 95\% CI.}

\item{check}{Whether to perform input check using \code{\link[=stopifnot]{stopifnot()}}.

Default is TRUE, use FALSE to save time.}
}
\value{
A named list with lower limit (LCI) and upper limit (UCI) of the
Wilson score confidence interval without continuity correction.
}
\description{
This function construct the (1 - alpha)\% Wilson score confidence interval
without continuity correction.

Either \code{x} or \code{n} can use vector as input to conduct vectorized operation for
multiple rates. Unequal length of \code{x} and \code{n} will throw an error (unless either
of the length equals 1).
}
\examples{
wilson_ci(10, 100)

# Several CIs
wilson_ci(c(10:15), 100)

# Several CIs, different x and n
wilson_ci(c(10, 20), c(100, 200))
}
\seealso{
\code{\link[=wilson_rd_ci]{wilson_rd_ci()}} for CI of rate difference.
}
